Create matplotlib sample using Db2 data and show various plot attributes (#59)
* Use standard folder structure for java and resources

With the standard structure, mvn package will gen the jar file with the correct files.

* Create matplotlib sample using Db2 data and show various plot attributes

* [machinelearning] add python3-pandas to list of packages to install

pandas is required for the machine learning demos (homeCredit and sklearnDemos)

* update requirements for machine learning

Update packages to install based on a clean 7.3 machine in order to run the matplotlib sample matplotwithdb2data.py

* [machinelearning] update matplotlib sample to work when tkinter is installed

set the matplotlib backend to write to a file so that when python3-tkinter is installed it still works.  Otherwise, a _tkinter.TclError would occur.